San Francisco 311 Case Index API
This includes the API to get and search the imported case data and an importer
to import the latest case data into the database.

The importer is scheduled to run at 7AM UTC. This is setup via whenever and
Cron. After deployment the 'whenever' command needs to be ran to update the
crontab file. Alternatively if capistrono is used to deploy the whenever
capistrono integration can be used.